requirements.txt
flask
pillow
waitress
gunicorn
